pyproject.toml
[project]
name = "intro-to-prefect"
description = "An introductory project to Prefect"
version = "0.1.0"
authors = [
    { name = "Bryan \"Beege\" Berry", email = "beege@codeseoul.org" }
]
dependencies = [
    "prefect",
    "httpx",
    "psycopg2-binary"
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "pytest"
]

[project.scripts]
pipeline = "pipeline.main:main"
